Handle None in tool/function arguments.
The normal OpenAI API returns an empty string when there are no function
arguments. The GitHub Copilot wrapper around the OpenAI API seems to
return None instead of an empty string when there are no arguments.

Treat None as an empty string. If we still have an empty string at the
point that we parse the JSON for the arguments, then use "{}" instead so
that we get the expected empty map.